Division better than for loop

+++ fs/zbd_zenfs.cc
@@ -165,7 +165,7 @@ inline IOStatus Zone::CheckRelease() {
 
 Zone *ZonedBlockDevice::GetIOZone(uint64_t offset) {
   uint64_t zone_sz=zbd_be_->GetZoneSize();
-  int zidx=offset/zone_sz - ZENFS_META_ZONES;
+  size_t zidx=offset/zone_sz - ZENFS_META_ZONES;
   if(io_zones.size()>zidx){
     return io_zones[zidx];
   }
